price page implementation
